Try to auth automatically to github.com from codespaces
This changes how the original gitconfig is included so that it doesn't
require changing any files inside the dotfiles workdir.

This sets up an askPass program that echoes GITHUB_TOKEN, along with a
default username (mine) for codespaces. My hope is that this works to
get me logged in. If it doesn't, I'll try defining my own credential
helper.